requirements.txt
numpy
Flask	
tensorflow
opencv-python	
Pillow			
requests